vendor/citrus: config: Lower the priority of vendor/citrus overlays
The priority order for overlays is PRODUCT > DEVICE > LOCAL
Having PRODUCT here was overriding all of the DEVICE_ specifics
mentioned in device repositories, which was not the intent.
vendor/cm is present as a set of sane defaults, but for the most
part these overlays are intended to be overridable by the device
if needed.

The logic is that PRODUCT_PACKAGE_OVERLAYS should be used by a
particular product while DEVICE_PACKAGE_OVERLAYS is used by
several products that share common device configs. Yeah, the naming
is counterintiuitive, but that's the implementation in
build/core/package_internal.mk

Partially addresses CYNGNOS-2229

Change-Id: Icfbc92593e37f3c00c3d26dd09e08d09267621ed
Signed-off-by: Adarsh-MR <adarshmr1998@gmail.com>